rearranging I
