+++ sbi/inference/trainers/marginal/__init__.py
@@ -1,1 +1,1 @@
-from sbi.inference.trainers.marginal.marginal_base import MarginalEstimator
+from sbi.inference.trainers.marginal.marginal_base import MarginalTrainer
