+++ sbi/neural_nets/factory.py
@@ -25,6 +25,7 @@ from sbi.neural_nets.net_builders.flow import (
     build_zuko_nsf,
     build_zuko_sospf,
     build_zuko_unaf,
+    build_zuko_unconditional_flow,
 )
 from sbi.neural_nets.net_builders.flowmatching_nets import (
     build_mlp_flowmatcher,
@@ -454,5 +455,65 @@ def posterior_score_nn(
             Callable: a ScoreEstimator object.
         """
         return build_score_estimator(batch_x=batch_theta, batch_y=batch_x, **kwargs)
+
+    return build_fn
+
+
+def marginal_nn(
+    model: str,
+    z_score_x: Optional[str] = "independent",
+    hidden_features: int = 50,
+    num_transforms: int = 5,
+    num_bins: int = 10,
+    num_components: int = 10,
+    **kwargs: Any,
+) -> Callable:
+    r"""
+    Returns a function that builds a density estimator for learning the marginal.
+
+    Args:
+        model: The type of density estimator that will be created. One of [`mdn`,
+            `made`, `maf`, `maf_rqs`, `nsf`].
+        z_score_x: Whether to z-score samples $x$ before passing them into
+            the network, can take one of the following:
+            - `none`, or None: do not z-score.
+            - `independent`: z-score each dimension independently.
+            - `structured`: treat dimensions as related, therefore compute mean and std
+            over the entire batch, instead of per-dimension. Should be used when each
+            sample is, for example, a time series or an image.
+        hidden_features: Number of hidden features.
+        num_transforms: Number of transforms when a flow is used. Only relevant if
+            density estimator is a normalizing flow (i.e. currently either a `maf` or a
+            `nsf`). Ignored if density estimator is a `mdn` or `made`.
+        num_bins: Number of bins used for the splines in `nsf`. Ignored if density
+            estimator not `nsf`.
+        num_components: Number of mixture components for a mixture of Gaussians.
+            Ignored if density estimator is not an mdn.
+        kwargs: additional custom arguments passed to downstream build functions.
+    """
+
+    kwargs = dict(
+        zip(
+            (
+                "z_score_x",
+                "hidden_features",
+                "num_transforms",
+                "num_bins",
+                "num_components",
+            ),
+            (
+                z_score_x,
+                hidden_features,
+                num_transforms,
+                num_bins,
+                num_components,
+            ),
+            strict=False,
+        ),
+        **kwargs,
+    )
+
+    def build_fn(batch_x):
+        return build_zuko_unconditional_flow(which_nf=model, batch_x=batch_x)
 
     return build_fn
